Modularise tests for easier use with wasi-common crate
